refactor(icons): extract materialExceptions into shared module
The Material-icon name table lived in two places: Icon.svelte (used by
@iconify/svelte rendering across the app) and maplibreSprites.ts (used
by the MapLibre pin-sprite pipeline). The maplibreSprites copy was the
older, shorter one — it never got PR #1004's six ic:outline fallback
entries, so /map-next pins for potted_plant / footprint / water_pump /
adult_content / raven / surgical fell through to the runtime cascade
instead of resolving directly.

Extract both the table and the `ic:outline-<name>` resolution into
`$lib/materialIcons.ts` as the single source of truth:
- Icon.svelte imports `resolveMaterialIcon`
- maplibreSprites.ts's `resolveIconifyName` wraps it (keeping only the
  map-specific question_mark → currency_bitcoin substitution)

/map-next now gets the six fallback entries for free, in sync with
prod /map. The runtime cascade in fetchIconInnerSvg stays as a safety
net for any future unknown tag value.

Branch rebased onto origin/main (picks up #1004 and #995) before this
change.
