+++ src/components/Icon.svelte
@@ -1,6 +1,8 @@
 <script lang="ts">
 import IconIconify from "@iconify/svelte";
 
+import { resolveMaterialIcon } from "$lib/materialIcons";
+
 export let w: string;
 export let h: string;
 let className: string | undefined = undefined;
@@ -9,33 +11,11 @@ export { className as class };
 export let icon: string;
 export let type: "material" | "fa" = "material";
 
-const materialExceptions: Record<string, string> = {
-	camping: "material-symbols:camping-rounded",
-	gate: "material-symbols:gate",
-	cooking: "material-symbols:cooking",
-	dentistry: "material-symbols:dentistry",
-	sauna: "material-symbols:sauna",
-	info_outline: "material-symbols:info-outline",
-	skull: "material-symbols:skull",
-	currency_bitcoin: "material-symbols:currency-bitcoin",
-	close_round: "ic:round-close",
-	my_location: "material-symbols:my-location-rounded",
-	bookmark_filled: "ic:baseline-bookmark-added",
-	account_circle_filled: "ic:baseline-account-circle",
-	// Missing from ic:outline — fall back to material-symbols
-	potted_plant: "material-symbols:potted-plant-outline",
-	footprint: "material-symbols:footprint-outline",
-	water_pump: "material-symbols:water-pump-outline",
-	adult_content: "material-symbols:explicit-outline",
-	raven: "material-symbols:raven-outline",
-	surgical: "material-symbols:surgical-outline",
-};
-
 const faBrandIcons = ["x-twitter", "instagram", "facebook", "twitter"];
 
 $: formattedIcon =
 	type === "material"
-		? materialExceptions[icon] || `ic:outline-${icon.replace(/_/g, "-")}`
+		? resolveMaterialIcon(icon)
 		: faBrandIcons.includes(icon)
 			? `fa6-brands:${icon}`
 			: `fa6-solid:${icon}`;
